+++ tweet-ui/src/main/java/com/twitter/sdk/android/tweetui/CompactTweetView.java
@@ -22,18 +22,14 @@ import android.content.Context;
 import android.os.Build;
 import android.util.AttributeSet;
 
-import com.squareup.picasso.Picasso;
 import com.twitter.sdk.android.core.models.MediaEntity;
 import com.twitter.sdk.android.core.models.Tweet;
-import com.twitter.sdk.android.tweetui.internal.util.AspectRatioImageView;
 
 public class CompactTweetView extends BaseTweetView {
     private static final String VIEW_TYPE_NAME = "compact";
     private static final double SQUARE_ASPECT_RATIO = 1.0;
     private static final double MAX_LANDSCAPE_ASPECT_RATIO = 3.0;
     private static final double MIN_LANDSCAPE_ASPECT_RATIO = 4.0 / 3.0;
-    static final double DEFAULT_ASPECT_RATIO = 16.0 / 9.0;
-    AspectRatioImageView aspectRatioPhotoView;
 
     public CompactTweetView(Context context, Tweet tweet) {
         super(context, tweet);
@@ -63,12 +59,6 @@ public class CompactTweetView extends BaseTweetView {
     }
 
     @Override
-    void findSubviews() {
-        super.findSubviews();
-        aspectRatioPhotoView = (AspectRatioImageView) findViewById(R.id.tw__tweet_media);
-    }
-
-    @Override
     void render() {
         super.render();
         // Redraw screen name on recycle
@@ -81,12 +71,9 @@ public class CompactTweetView extends BaseTweetView {
      * @param photoEntity the first
      * @return the target image and bitmap width to height aspect ratio
      */
-    static double getAspectRatio(MediaEntity photoEntity) {
-        if (photoEntity == null || photoEntity.sizes == null || photoEntity.sizes.medium == null ||
-                photoEntity.sizes.medium.w == 0 || photoEntity.sizes.medium.h == 0) {
-            return DEFAULT_ASPECT_RATIO;
-        }
-        final double ratio = (double) photoEntity.sizes.medium.w / photoEntity.sizes.medium.h;
+    @Override
+    protected double getAspectRatio(MediaEntity photoEntity) {
+        final double ratio = super.getAspectRatio(photoEntity);
         if (ratio <= SQUARE_ASPECT_RATIO) {
             // portrait (tall) photos should be cropped to be square aspect ratio
             return SQUARE_ASPECT_RATIO;
@@ -100,28 +87,6 @@ public class CompactTweetView extends BaseTweetView {
             // landscape photos between 3:1 to 4:3 present the original width to height ratio
             return ratio;
         }
-    }
-
-    @Override
-    protected void setTweetPhoto(final MediaEntity photoEntity) {
-        final Picasso imageLoader = dependencyProvider.getImageLoader();
-        if (imageLoader == null) return;
-
-        // set aspect ratio the AspectRatioImageView will respect when it is measured
-        final double aspectRatio = getAspectRatio(photoEntity);
-        aspectRatioPhotoView.setAspectRatio(aspectRatio);
-
-        // Picasso fit is a deferred call to resize(w,h) which waits until the target has a
-        // non-zero width or height and resizes the bitmap to the target's width and height.
-        // For recycled targets, which already have a width and (stale) height, reset the size
-        // target to zero so Picasso fit works correctly.
-        aspectRatioPhotoView.resetSize();
-
-        imageLoader.load(photoEntity.mediaUrlHttps)
-                .placeholder(mediaBg)
-                .fit()
-                .centerCrop()
-                .into(mediaPhotoView, new PicassoCallback());
     }
 
     @Override